requirement.txt
django==1.5
mysql-python
python-openid
django-openid-auth
django_socketio
